add full service control and order control and etc..

+++ app/Http/Controllers/Auth/ServiceController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Auth;
 use App\Http\Controllers\Controller;
 use App\Models\Service;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class ServiceController extends Controller
 {
@@ -13,8 +14,14 @@ class ServiceController extends Controller
      */
     public function index()
     {
-        $services = Service::paginate(10);
-        return view('admin.services', ['services' => $services]);
+        if (Auth::user()->isAdmin()) {
+            $services = Service::paginate(10);
+            return view('admin.services', ['services' => $services]);
+        }
+        if (Auth::user()->isUser()) {
+            $services = Service::paginate(10);
+            return view('auth.services', ['services' => $services]);
+        }
     }
 
     /**
@@ -22,7 +29,10 @@ class ServiceController extends Controller
      */
     public function create()
     {
-        return view('admin.serviceCreate');
+        if (Auth::user()->isAdmin()) {
+            return view('admin.serviceCreate');
+        }
+        return redirect()->route('welcome');
     }
 
     /**
@@ -30,16 +40,19 @@ class ServiceController extends Controller
      */
     public function store(Request $request)
     {
-        $request->validate([
-            'name' => 'required|min:5',
-            'discriprion' => 'required'
-        ]);
-        $service = Service::create([
-            'name' => $request->name,
-            'discriprion' => $request->discriprion
-        ]);
-        $service->save();
-        return redirect()->route('service.index')->with('status', 'Услуга ' . $request->name . ' успешно добавлена');
+        if (Auth::user()->isAdmin()) {
+            $request->validate([
+                'name' => 'required|min:5',
+                'discriprion' => 'required'
+            ]);
+            $service = Service::create([
+                'name' => $request->name,
+                'discriprion' => $request->discriprion
+            ]);
+            $service->save();
+            return redirect()->route('service.index')->with('status', 'Услуга ' . $request->name . ' успешно добавлена');
+        }
+        return redirect()->route('welcome');
     }
 
     /**
@@ -47,8 +60,15 @@ class ServiceController extends Controller
      */
     public function show(string $id)
     {
-        $service = Service::find($id);
-        return view('admin.service', ['service' => $service]);
+        if (Auth::user()->isAdmin()) {
+            $service = Service::find($id);
+            return view('admin.service', ['service' => $service]);
+        }
+        if (Auth::user()->isUser()) {
+            $service = Service::find($id);
+            return view('auth.service', ['service' => $service]);
+        }
+        return redirect()->route('welcome');
     }
 
     /**
@@ -56,8 +76,11 @@ class ServiceController extends Controller
      */
     public function edit(string $id)
     {
-        $service = Service::find($id);
-        return view('admin.serviceEdit', ['service' => $service]);
+        if (Auth::user()->isAdmin()) {
+            $service = Service::find($id);
+            return view('admin.serviceEdit', ['service' => $service]);
+        }
+        return redirect()->route('welcome');
     }
 
     /**
@@ -65,16 +88,19 @@ class ServiceController extends Controller
      */
     public function update(Request $request, string $id)
     {
-        $service = Service::find($id);
-        $request->validate([
-            'name' => 'required',
-            'discriprion' => 'required'
-        ]);
-        $service->update([
-            'name' => $request->name,
-            'discriprion' => $request->discriprion
-        ]);
-        return redirect()->route('service.index')->with('status', 'Услуга №' . $service->id . ' успешно изменена');
+        if (Auth::user()->isAdmin()) {
+            $service = Service::find($id);
+            $request->validate([
+                'name' => 'required',
+                'discriprion' => 'required'
+            ]);
+            $service->update([
+                'name' => $request->name,
+                'discriprion' => $request->discriprion
+            ]);
+            return redirect()->route('service.index')->with('status', 'Услуга №' . $service->id . ' успешно изменена');
+        }
+        return redirect()->route('welcome');
     }
 
     /**
@@ -82,8 +108,11 @@ class ServiceController extends Controller
      */
     public function destroy(string $id)
     {
-        $service = Service::find($id);
-        $service->delete();
-        return redirect()->route('service.index')->with('status', 'Услуга ' . $service->name . ' удалена');
+        if (Auth::user()->isAdmin()) {
+            $service = Service::find($id);
+            $service->delete();
+            return redirect()->route('service.index')->with('status', 'Услуга ' . $service->name . ' удалена');
+        }
+        return redirect()->route('welcome');
     }
 }
